+++ lib/shelf_plus.dart
@@ -3,12 +3,13 @@
 library shelf_plus;
 
 export 'package:shelf/shelf.dart';
-export 'package:shelf_static/shelf_static.dart';
 export 'package:shelf_router/shelf_router.dart';
+export 'package:shelf_static/shelf_static.dart';
 
+export 'src/request_body_accessor.dart';
 export 'src/router_plus/cascade_helper.dart';
-export 'src/router_plus/middlewares.dart';
 export 'src/router_plus/middleware_extension.dart';
+export 'src/router_plus/middlewares.dart';
 export 'src/router_plus/response_handler/response_handler.dart';
 export 'src/router_plus/router_plus.dart';
 export 'src/router_plus/shelf_request_extension.dart';