+++ tool/templates/code/other_code.dart
@@ -4,8 +4,12 @@ class Person {
 
   Person({required this.name, required this.age});
 
-  // created with tools like @jsonSerializable
+  // created with tools like json_serializable package
   Map<String, dynamic> toJson() {
     return {'name': name, 'age': age};
+  }
+
+  static Person fromJson(Map<String, dynamic> json) {
+    return Person(name: '', age: 0);
   }
 }