+++ tool/templates/code/request_body_deserialize1.dart
@@ -1,0 +1,18 @@
+import 'package:shelf_plus/shelf_plus.dart';
+
+import 'other_code.dart';
+
+void main() => shelfRun(init);
+
+Handler init() {
+  var app = Router().plus;
+
+  app.post('/json', (Request request) async {
+    //@start
+    var person = await request.body.as(Person.fromJson);
+    //@end
+    return 'You send me: ${person.name}';
+  });
+
+  return app;
+}