+++ tool/templates/code/request_body_deserialize2.dart
@@ -1,0 +1,10 @@
+class Person {
+  final String name;
+
+  Person({required this.name});
+
+  // created with tools like json_serializable package
+  static Person fromJson(Map<String, dynamic> json) {
+    return Person(name: json['name']);
+  }
+}