+++ tool/templates/code/request_body_ext_model1.dart
@@ -1,0 +1,9 @@
+import 'package:shelf_plus/shelf_plus.dart';
+
+import 'other_code.dart';
+
+//@start
+extension PersonAccessor on RequestBodyAccessor {
+  Future<Person> get asPerson async => Person.fromJson(await asJson);
+}
+//@end