+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## next
+
+- **New:** Request Body Accessor
+
 ## 0.0.1
 
 - Initial version